Added cookieValidation.

+++ server/src/java/server/resources/ReplacementResource.java
@@ -20,6 +20,7 @@ import org.hibernate.criterion.Restrictions;
 
 import server.entities.Replacement;
 import server.entities.ReplacementDeserializer;
+import server.operations.CookieHandler;
 import server.persistence.HibernateUtil;
 
 import com.google.gson.Gson;
@@ -39,46 +40,50 @@ public class ReplacementResource {
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public final String changeReplacementJSON(@CookieParam(value = "NSA-Cookie") final String nsaCookie, final String replacementJSON) {
-		// TODO: check nsaCookie
-
-		final GsonBuilder gson = new GsonBuilder();
-		gson.registerTypeAdapter(Replacement.class, new ReplacementDeserializer());
-		final Replacement replacement = gson.create().fromJson(replacementJSON, Replacement.class);
-		final EntityManager entityManager = HibernateUtil.getEntityManager();
-		entityManager.getTransaction().begin();
-		HibernateUtil.getEntityManager().persist(replacement);
-		entityManager.getTransaction().commit();
-		return new Gson().toJson(replacement);
+		// TODO: return null okay?
+		if (new CookieHandler().validateCookie(nsaCookie)) {
+			final GsonBuilder gson = new GsonBuilder();
+			gson.registerTypeAdapter(Replacement.class, new ReplacementDeserializer());
+			final Replacement replacement = gson.create().fromJson(replacementJSON, Replacement.class);
+			final EntityManager entityManager = HibernateUtil.getEntityManager();
+			entityManager.getTransaction().begin();
+			HibernateUtil.getEntityManager().persist(replacement);
+			entityManager.getTransaction().commit();
+			return new Gson().toJson(replacement);
+		}
+		return null;
 	}
 
 	@POST
 	@Consumes(MediaType.APPLICATION_JSON)
 	@Produces(MediaType.APPLICATION_JSON)
 	public final String addReplacementJSON(@CookieParam(value = "NSA-Cookie") final String nsaCookie, final String replacementJSON) {
-		// TODO: check nsaCookie
+		// TODO: return null okay?
+		if (new CookieHandler().validateCookie(nsaCookie)) {
+			final GsonBuilder gson = new GsonBuilder();
+			gson.registerTypeAdapter(Replacement.class, new ReplacementDeserializer());
+			final Replacement replacement = gson.create().fromJson(replacementJSON, Replacement.class);
 
-		final GsonBuilder gson = new GsonBuilder();
-		gson.registerTypeAdapter(Replacement.class, new ReplacementDeserializer());
-		final Replacement replacement = gson.create().fromJson(replacementJSON, Replacement.class);
-
-		return new Gson().toJson(this.addReplacement(replacement));
+			return new Gson().toJson(this.addReplacement(replacement));
+		}
+		return null;
 	}
 
 	@DELETE
-	public final boolean deleteReplacement(@CookieParam(value = "NSA-Cookie") final String nsaCookie,
-			@QueryParam("id") final int replacementId) {
-		// TODO: check nsaCookie
+	public final boolean deleteReplacement(@CookieParam(value = "NSA-Cookie") final String nsaCookie, @QueryParam("id") final int replacementId) {
+		if (new CookieHandler().validateCookie(nsaCookie)) {
+			final EntityManager entityManager = HibernateUtil.getEntityManager();
+			final Replacement replacement = entityManager.find(Replacement.class, replacementId);
+			if (replacement == null) {
+				return false;
+			}
 
-		final EntityManager entityManager = HibernateUtil.getEntityManager();
-		final Replacement replacement = entityManager.find(Replacement.class, replacementId);
-		if (replacement == null) {
-			return false;
+			entityManager.getTransaction().begin();
+			entityManager.remove(replacement);
+			entityManager.getTransaction().commit();
+			return true;
 		}
-		
-		entityManager.getTransaction().begin();
-		entityManager.remove(replacement);
-		entityManager.getTransaction().commit();
-		return true;
+		return false;
 	}
 
 	private Replacement addReplacement(final Replacement replacement) {
@@ -93,7 +98,7 @@ public class ReplacementResource {
 	private List<?> getReplacements(final int teacherId, final int formId, final int roomId, final String week) {
 		final Session session = HibernateUtil.getEntityManager().unwrap(Session.class);
 		final Criteria criteria = session.createCriteria(Replacement.class);
-		
+
 		if (teacherId != 0) {
 			criteria.add(Restrictions.eq("teacher.id", teacherId));
 		}
@@ -106,7 +111,7 @@ public class ReplacementResource {
 		if (week != null) {
 			criteria.add(Restrictions.eq("week", week));
 		}
-		
+
 		return criteria.list();
 	}
 }
